build: build character appearances query logic
The logic regarding API requests to the Fandom API for character appearances and how the return is processed was built. It was necessary to refactor the service/client structure so that the client's (FandomApiClient) sole responsibility is to send API requests while a FandomQueryService processes the payload content and 'repackages' the data for SearchService's consumption.