build: apply the character appearances query logic
The external API search for character issues and its processing is functional, particularly for DC comics characters. However, the PublisherSearchService interface will most likely be turned into an abstract class. Also changed the signatures of a number of methods so that they'll receive a PublisherEnum as an argument rather than PublisherEndpointEnum