feat: add character appearances persistence mechanism
When a query is made, the application's very first action is to look for the up-to-date data on the queried character's appearances. It will only rely on the external API in case the data isn't recorded in the database or if the available data is considered outdated. It was established that the available data must be updated in case the last update on the character version's appearances was made over 3 days prior to the query