perf: change the character appearance query procedures
The way it was before worked, but the request duration dramatically increased the more appearances a character has. It used to take around 6 minutes for the application to fetch all the necessary data about Raven's 344 appearances from the Fandom API, for example, but after the changes, that time was trimmed down to 21 seconds. Once persisted, the data will be retrieved in a matter of milliseconds (around 120 ms). But there's still room for improvement