feat: build disambiguation persistence logic
In SearchService, if the queried term doesn't return character versions already listed in the database, then it should rely on a Fandom query to obtain such data, assuming the term is a valid character alias. If no data is found, it will ultimately lead to an error message returned to the client; if data is found, the application's database will be updated. Some tweaking was necessary on some of these classes, related to exception handling. The database update part has yet to be completed