feat: finish disambiguation persistence logic
When the queried term is a valid character name and the database query yields results, a response will promptly be sent to the client containing these results. If no data is found, a fandom query is made, and if this query yields results, they will be recorded in the database for future queries, and the results will be sent as a response to the client as intended. If the fandom query doesn't yield results, it means the term isn't a valid character alias. This logic has been implemented