refactor: distinguish FandomQueryService implementations by publishers
Upon further study on how the disambiguation pages are structured in a Marvel Fandom API response, it was noticed that they are wildly different from DC's API responses, which affects data parsing. It was necessary to turn FandomQueryService into an abstract class containing all the methods that were already in use (outside of the getDisambiguation() method, all the other methods were implemented the same way), but getDisambiguation() is declared as an abstract method to force each publisher service to implement it in its own particular way. If by any chance another possible publisher Fandom API demands that the other methods be implemented in a different way, its service can always override the parent's method